--- env.go
@@ -1,5 +1,0 @@
-package geocodio
-
-const (
-	EnvGeocodioAPIKey = "GEOCODIO_API_KEY"
-)
