--- location.go
@@ -1,6 +1,0 @@
-package geocodio
-
-type Location struct {
-	Latitude  float64 `json:"lat"`
-	Longitude float64 `json:"lng"`
-}
